+++ src/types/OfferItems.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.18;
 
-import "contracts/types/OfferItem.sol";
+import "./OfferItem.sol";
 
 // @dev Struct representing multiple offer items. We store the count for decoding purposes
 struct OfferItems {
-    uint16 chainId;
+    uint256 chainId;
     OfferItem[] items;
 }
